requirements.txt
pytest==8.0.2
playwright==1.41.2
pytest-playwright==0.4.4
allure-pytest==2.13.2
pytest-xdist==3.8.0
